+++ scholar/model/__init__.py
@@ -1,5 +1,5 @@
 from .nn import MLPLM
-from .nn import MyLM
-from .nn import ABPCNLM
-from .nn import MinervaConfig, MinervaNLM
+from .abpc import ABPCNLM
+from .minerva import MinervaConfig, MinervaNLM
 from .transformer import TransformerLM
+from .streamformer import StreamformerLM
