+++ scholar/model/streamformer.py
@@ -2,36 +2,47 @@ import math
 import torch
 from torch.nn import Module, Linear, Dropout, LayerNorm
 from torch.cuda.amp import autocast
-from .nn import Sequential, Embedding, MLP, LanguageModel, ResidualDropoutLayerNorm
+from .nn import Sequential, Embedding, MLP, LanguageModel, Lambda
+
+
+class ResidualLayerNorm(Module):
+    def __init__(self, layer, d_model):
+        super().__init__()
+        self.d_model = d_model
+
+        self.layer = layer
+        self.layernorm = LayerNorm(d_model)
+
+    def forward(self, x):
+        assert x.shape[-1] == self.d_model, f"{x.shape[-1]} != {self.d_model}"
+        return self.layernorm(x+self.layer(x))
 
 
 class Mask(Module):
-    def __init__(self, mask="none"):
+    def __init__(self, mode="none"):
         super().__init__()
-        self.mask = mask
+        self.mode = mode
 
     def forward(self, x):
         n, device = x.shape[-1], x.device
-        if self.mask == "none":
+        if self.mode == "none":
             return x
-        elif self.mask == "causal":
+        elif self.mode == "causal":
             return x+(1-1/torch.tril(torch.ones((n,n),device=device)))
 
 
 class Attn(Module):
-    def __init__(self, d_model, d_k, d_v, n_heads, p_dropout, mask="none"):
+    def __init__(self, d_model, d_k, d_v, n_heads):
         super().__init__()
         self.d_model = d_model
         self.d_k = d_k
         self.d_v = d_v
         self.n_heads = n_heads
-        self.p_dropout = p_dropout
 
         self.query_proj = Linear(d_model, d_k*n_heads)
         self.key_proj = Linear(d_model, d_k*n_heads)
         self.value_proj = Linear(d_model, d_v*n_heads)
-        self.mask = Mask(mask=mask)
-        self.dropout = Dropout(p_dropout)
+        self.mask = Mask(mode="none")
         self.softmax = torch.nn.Softmax(dim=-1)
         self.linear = Linear(d_v*n_heads, d_model, bias=False)
 
@@ -42,24 +53,20 @@ class Attn(Module):
         merge_heads = lambda x: x.transpose(-2,-3).contiguous().view(x.shape[:-3]+(n_ctx,self.d_v*self.n_heads))
         (Q, K, V) = map(split_heads,(self.query_proj(x),self.key_proj(x),self.value_proj(x)))
         QKT = torch.matmul(Q/math.sqrt(self.d_k),K.transpose(-1,-2))
-        return self.linear(merge_heads(self.dropout(self.softmax(self.mask(QKT)))@V))
+        return self.linear(merge_heads(self.softmax(self.mask(QKT))@V))
 
 
-class TransformerLayer(Module):
-    def __init__(self, d_model, d_k, d_v, n_heads, d_hidden, p_dropout_attn_mat, p_dropout_attn_out, p_dropout_mlp, mask="none"):
+class StreamformerLayer(Module):
+    def __init__(self, d_model, d_k, d_v, n_heads, d_hidden):
         super().__init__()
         self.d_model = d_model
         self.d_k = d_k
         self.d_v = d_v
         self.n_heads = n_heads
         self.d_hidden = d_hidden
-        self.p_dropout_attn_mat = p_dropout_attn_mat
-        self.p_dropout_attn_out = p_dropout_attn_out
-        self.p_dropout_mlp = p_dropout_mlp
-        self.mask = mask
 
-        self.attn = ResidualDropoutLayerNorm(Attn(d_model, d_k, d_v, n_heads, p_dropout_attn_mat, mask), d_model, p_dropout_attn_out)
-        self.mlp = ResidualDropoutLayerNorm(MLP(d_model, d_hidden, 'GELU', d_model), d_model, p_dropout_mlp)
+        self.attn = ResidualLayerNorm(Attn(d_model, d_k, d_v, n_heads), d_model)
+        self.mlp = ResidualLayerNorm(MLP(d_model, d_hidden, 'GELU', d_model), d_model)
 
     def forward(self, x):
         return self.mlp(self.attn(x))
@@ -78,7 +85,7 @@ class PositionalEncoding(Module):
         return x + self.weight[:n_ctx]
 
 
-class TransformerLM(Module):
+class StreamformerLM(Module):
     def __init__(self,
                  n_vocab_in,
                  n_vocab_out,
@@ -89,12 +96,6 @@ class TransformerLM(Module):
                  n_heads,
                  d_hidden,
                  n_layers,
-                 p_dropout_embedding=0.0,
-                 p_dropout_attn_mat=0.0,
-                 p_dropout_attn_out=0.0,
-                 p_dropout_mlp=0.0,
-                 mode='shift',
-                 mask='causal',
                  autocast_enabled=None):
         super().__init__()
         self.n_vocab_in = n_vocab_in
@@ -106,12 +107,7 @@ class TransformerLM(Module):
         self.n_heads = n_heads
         self.d_hidden = d_hidden
         self.n_layers = n_layers
-        self.p_dropout_embedding = p_dropout_embedding
-        self.p_dropout_attn_mat = p_dropout_attn_mat
-        self.p_dropout_attn_out = p_dropout_attn_out
-        self.p_dropout_mlp = p_dropout_mlp
         self.autocast_enabled = autocast_enabled or False
-        self.mask = mask
         self.language_model = (
             LanguageModel(
                 n_vocab_out=n_vocab_out,
@@ -119,10 +115,9 @@ class TransformerLM(Module):
                 module=
                     Sequential(
                         Embedding(n_vocab_in, d_model),
-                        Dropout(p_dropout_embedding),
                         PositionalEncoding(n_ctx, d_model),
                         Sequential(
-                            TransformerLayer(d_model, d_k, d_v, n_heads, d_hidden, p_dropout_attn_mat, p_dropout_attn_out, p_dropout_mlp, mask) for _ in range(n_layers)),
+                            StreamformerLayer(d_model, d_k, d_v, n_heads, d_hidden) for _ in range(n_layers)),
                         Linear(d_model, n_vocab_out))))
 
     def forward(self, x):
